got rid of angular
I decided to go with knockout.js since it just does data bindings
(which is all we really need).  We want to keep the tech simple so that
we can focus on learning dev practices and not new frameworks.  Let me
know if anyone WANTS to use angular and maybe we can bring it back.